backend/requirements.txt
fastapi
uvicorn
python-dotenv
# pydantic
langchain-google-genai
langchain_openai
langchain_core
langchain_community

sqlalchemy
databases

stability-sdk